Moved JS vars from admin to front handler
